+++ src/battle_anim_effects_3.c
@@ -1441,7 +1441,7 @@ void AnimTask_SetPsychicBackground(u8 taskId)
 
 static void SetPsychicBackground_Step(u8 taskId)
 {
-    int i;
+    u32 i;
     u16 lastColor;
 
     if (++gTasks[taskId].data[5] == 4)
@@ -1466,7 +1466,7 @@ void AnimTask_FadeScreenToWhite(u8 taskId)
 
 static void FadeScreenToWhite_Step(u8 taskId)
 {
-    int i;
+    u32 i;
     u16 lastColor;
 
     if (++gTasks[taskId].data[5] == 4)
@@ -1949,7 +1949,7 @@ static void TormentAttacker_Step(u8 taskId)
 {
     int var0, var1;
     s16 x, y;
-    u16 i, j;
+    u32 i, j;
     u8 spriteId;
     struct Task *task = &gTasks[taskId];
 
@@ -3798,7 +3798,7 @@ static void AnimTask_SquishAndSweatDroplets_Step(u8 taskId)
 
 static void CreateSweatDroplets(u8 taskId, bool8 lowerDroplets)
 {
-    u8 i;
+    u32 i;
     s8 xOffset, yOffset;
     struct Task *task;
     s16 xCoords[4];
@@ -3999,7 +3999,7 @@ void AnimTask_GlareEyeDots(u8 taskId)
 
 static void AnimTask_GlareEyeDots_Step(u8 taskId)
 {
-    u8 i;
+    u32 i;
     s16 x, y;
     struct Task *task = &gTasks[taskId];
 
@@ -4737,7 +4737,7 @@ static void AnimMeteorMashStar(struct Sprite *sprite)
 
 void AnimTask_MonToSubstitute(u8 taskId)
 {
-    int i;
+    u32 i;
     u8 spriteId = GetAnimBattlerSpriteId(ANIM_ATTACKER);
 
     if (gTasks[taskId].data[0] == 0)
